backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
requests==2.31.0
Werkzeug==2.3.7
Gunicorn==21.2.0
